Rework the detection preview.
We do not recompute probabilities when the frame and the ROI do
not change. It saves a lot of time when adjusting the probability
threshold, because ilastik takes a lot of time to launch everytime.
Here we used the same strategy that for the Weka detector: we
simply store the last probability map and check whether we need
to recompute it next time the user presses the 'Preview' button.
Nowe I have to check that it also works for 3D.